refactor: use array for passing arguments in scripts
Use array and array expansion for passing arguments to commands
inside the `sh` files. Some newer scripts already started with this
practice and now it is applied across the older ones as well.
This avoids using `eval` since it is not necessary making code safe,
predictable, cleaner and maintainable.
Affected commands and jobs:
- `analyze_code`
- `analyze_code_diff`
- `analyze_code_full`
- `detect_secrets`
- `detect_secrets_dir`
- `detect_secrets_git`